[NFC] Basic: Extract feature API definitions into their own file
